requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.0.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
discord.py==2.1.1
distro==1.7.0
distro-info===1.1build1
frozenlist==1.3.3
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
openai==0.26.5
Pillow==9.4.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0
python-dotenv==0.21.1
PyYAML==5.4.1
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tqdm==4.64.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.6
yarl==1.8.2
zipp==1.0.0
